requirements.txt
requests>=2.25.1
PyJWT>=2.0.0
pydantic>=1.10.2
urllib3>=1.26.0